Add function to check if file was already submitted so it wont be submitted again

+++ src/csv_writer.py
@@ -1,9 +1,22 @@
 import csv
+from os import path
+
+# check if sample is in csv log file
+def check_recorded(log_f, log_dir, f):
+    with open(log_dir+log_f, encoding="utf-8") as csvfile:
+        reader = csv.reader(csvfile, delimiter=',', quotechar='"')
+        header = next(reader)
+        for row in reader:
+            row = dict(zip(header,row))
+            if row['Filename'] == path.basename(f):
+                return True
+        return False
 
 # return md5 hash of the submited sample
 def get_hash(sample_id: str, client) -> str:
     return client.overview_report(sample_id)['sample']['md5']
 
+# create csv file name
 def create_file_name(directory):
     log_f = directory.replace('/','_')
     if log_f.endswith('_'):
@@ -12,12 +25,13 @@ def create_file_name(directory):
         log_f = log_f + ".csv"
     return log_f
 
-
+# write csv header if csv file doest exists
 def write_header(file, log_dir):
-    header = ['Filename', 'Sample_id', 'mb5_hash']
-    with open(log_dir+file, 'w') as f:
-        writer = csv.writer(f)
-        writer.writerow(header)
+    if not path.exists(log_dir+file):
+        header = ['Filename', 'Sample_id', 'md5_hash']
+        with open(log_dir+file, 'w') as f:
+            writer = csv.writer(f)
+            writer.writerow(header)
 
 # log sample id and from witch directory its from
 def log(sample_id : str, filename : str, log_f : str, client, log_dir):
